requirements.txt
# Core
python-dotenv
requests

# AI
openai

# Vector DB
faiss-cpu
numpy

# Web UI
streamlit